refactor(mapgen): purify live diagnostic ownership
